+++ .github/workflows/check-projects.yml
@@ -38,8 +38,8 @@ jobs:
       - name: Advanced-Audio-Driver
         run: cd Advanced-Audio-Driver && ./gradlew app:assembleRelease && cd ..
 
-      - name: Custom-Video-Driver
-        run: cd Custom-Video-Driver && ./gradlew app:assembleRelease && cd ..
+      - name: Basic-Video-Driver
+        run: cd Basic-Video-Driver && ./gradlew app:assembleRelease && cd ..
 
       - name: Basic-Video-Capturer
         run: cd Basic-Video-Capturer && ./gradlew app:assembleRelease && cd ..
