+++ Basic-Video-Driver/app/build.gradle
@@ -11,7 +11,7 @@ android {
     buildToolsVersion extBuildToolsVersion
 
     defaultConfig {
-        applicationId "com.tokbox.sample.customvideodriver"
+        applicationId "com.tokbox.sample.basicvideodriver"
         minSdkVersion 21
         targetSdkVersion extTargetSdkVersion
         versionCode extVersionCode
