+++ Basic-Video-Driver/app/src/main/res/values-w820dp/dimens.xml
