+++ Basic-Video-Driver/app/src/main/res/values/strings.xml
@@ -1,4 +1,4 @@
 <resources>
-    <string name="app_name">Custom-Video-Driver</string>
+    <string name="app_name">Basic-Video-Driver</string>
     <string name="rationale_video_app">This app needs access to your camera and mic so you can perform video calls</string>
 </resources>
