+++ Basic-Video-Driver/app/src/main/res/values/styles.xml
